+++ src/Test/Fixtures/TrailFixture.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Hirtz\Skeleton\Test\Fixtures;
 
 use Hirtz\Skeleton\Models\Trail;
-use yii\test\ActiveFixture;
 
 class TrailFixture extends ActiveFixture
 {